refactor(example): co-location bridge code
