+++ example/native/App.tsx
@@ -7,8 +7,25 @@
 
 import React from "react";
 import { Button, SafeAreaView } from "react-native";
-import type { RNBridgeWebView } from "@rnbridge/native";
-import { WebView, WebMethod } from "./src/bridge";
+import { bridge, createWebView, type RNBridgeWebView } from "@rnbridge/native";
+import InAppBrowser from "react-native-inappbrowser-reborn";
+import { WebBridge } from "@rnbridge/example-web";
+
+export const appBridge = bridge({
+  getMessage: () => {
+    return "I'm from native" as const;
+  },
+  openInAppBrowser: async (url: string) => {
+    if (await InAppBrowser.isAvailable()) {
+      await InAppBrowser.open(url);
+    }
+  },
+});
+
+export const { WebView, WebMethod } = createWebView<WebBridge>({
+  bridge: appBridge,
+  debug: true,
+});
 
 function App(): JSX.Element {
   const webviewRef = React.useRef<RNBridgeWebView>(null);
